+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+- Added workbench iife to load/toggle concise-syntax css using a manual status bar icon.
 - Bug editorBracketHighlight skips as concise target. Bug this extension overrides the user's theme.
 - Added injection.json: match closing tags. Added tokens.json: hide redundant syntax.
 - Init